update runtest script
