--- test/local_runtests.jl
@@ -1,4 +1,0 @@
-include("../src/DecisionTree.jl")
-using Main.DecisionTree
-
-include("test_suites.jl")
